+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/utils/WidgetFactory.java
@@ -278,6 +278,16 @@ public class WidgetFactory {
     return controller.getResult();
   }
 
+  public static ConfirmationResult showAlertOptionWithMandatoryCheckbox(Stage owner, String msg, String altOptionText, String okText, String help1, String help2, String checkBoxText) {
+    Stage stage = createDialogStage(ConfirmationDialogWithCheckboxController.class, owner, "Information", "dialog-alert-option-with-checkbox.fxml");
+    ConfirmationDialogWithCheckboxController controller = (ConfirmationDialogWithCheckboxController) stage.getUserData();
+    controller.hideCancel();
+    controller.initDialog(stage, altOptionText, okText, msg, help1, help2, checkBoxText);
+    controller.setCheckboxMandatory();
+    stage.showAndWait();
+    return controller.getResult();
+  }
+
   public static String showInputDialog(Stage owner, String dialogTitle, String innerTitle, String description, String helpText, String defaultValue) {
     Stage stage = createDialogStage(InputDialogController.class, owner, dialogTitle, "dialog-input.fxml");
     InputDialogController controller = (InputDialogController) stage.getUserData();