+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/GamesResource.java
@@ -2,7 +2,6 @@ package de.mephisto.vpin.server.games;
 
 import de.mephisto.vpin.commons.utils.FileUtils;
 import de.mephisto.vpin.restclient.descriptors.DeleteDescriptor;
-import de.mephisto.vpin.restclient.descriptors.ResetHighscoreDescriptor;
 import de.mephisto.vpin.restclient.descriptors.TableUploadDescriptor;
 import de.mephisto.vpin.restclient.popper.TableDetails;
 import de.mephisto.vpin.restclient.representations.ValidationState;
@@ -16,7 +15,6 @@ import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
@@ -24,6 +22,7 @@ import org.springframework.web.server.ResponseStatusException;
 import java.io.File;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
 import static de.mephisto.vpin.server.VPinStudioServer.API_SEGMENT;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
@@ -102,9 +101,9 @@ public class GamesResource {
     return gameService.deleteGame(descriptor);
   }
 
-  @PostMapping("/reset")
-  public boolean reset(@RequestBody ResetHighscoreDescriptor descriptor) {
-    return gameService.resetGame(descriptor.getGameId());
+  @DeleteMapping("/reset/{gameId}")
+  public boolean reset(@PathVariable("gameId") int gameId) {
+    return gameService.resetGame(gameId);
   }
 
   @PostMapping("/save")